feat: 支持智能机器人 WebSocket 长连接与审批回调
新增智能机器人 WebSocket 长连接（SmartRobotWs），包括 ISmartRobotWsClient 接口与 SmartRobotWsClient 实现，支持消息/事件实时回调、主动推送、分片上传素材、文档工具调用等能力。WecomOptions 新增 BotId/BotSecret/BotWsUrl 配置，WecomClient/DI 支持注入 WebSocket 客户端。CallbackReplyBuilder 新增 BuildMarkdown/BuildTemplateCard，支持被动回复 Markdown/模板卡片。

新增审批回调（open_approval_change）强类型解析，补全 CallbackApprovalEvent 及相关模型。IApprovalService/ApprovalService 新增链式审批申请构建器（ApplyEventBuilder）和自动分页获取审批单号方法。ApplyContentValue 及相关模型补全假勤、公式、关联审批等控件，细化字段。

优化 CallbackMessageBase 支持 attachment 消息、enter_chat 事件解析。SmartRobotService/ISmartRobotService 接口升级，支持智能表格自动化群聊管理。WecomServiceCollectionExtensions 注册 ISmartRobotWsClient。完善 README，补充 API/WS 用法、通讯录变更/审批/异步任务回调说明与示例。

本次更新大幅提升 SDK 对企业微信智能机器人（API/WS）、审批流、通讯录变更等场景的易用性和完备性，适合深度集成自动化与智能机器人开发者使用。

+++ Models/Wecom/Approval/ApplyLocationValue.cs
@@ -18,7 +18,7 @@ public class ApplyLocationValue
     /// <summary>位置地址</summary>
     [JsonPropertyName("address")] public string? Address { get; set; }
 
-    /// <summary>精确到多少米</summary>
-    [JsonPropertyName("time")] public int? Time { get; set; }
+    /// <summary>选择地点的时间（Unix 时间戳）</summary>
+    [JsonPropertyName("time")] public long? Time { get; set; }
 }
